+++ app/src/main/java/ly/count/android/demo/ActivityExampleDeviceId.java
@@ -36,22 +36,4 @@ public class ActivityExampleDeviceId extends AppCompatActivity {
         //give all consents after entering temporary device ID mode (that removes all consent)
         Countly.sharedInstance().consent().giveConsentAll();
     }
-
-    @Override
-    public void onStart() {
-        super.onStart();
-        Countly.sharedInstance().onStart(this);
-    }
-
-    @Override
-    public void onStop() {
-        Countly.sharedInstance().onStop();
-        super.onStop();
-    }
-
-    @Override
-    public void onConfigurationChanged(Configuration newConfig) {
-        super.onConfigurationChanged(newConfig);
-        Countly.sharedInstance().onConfigurationChanged(newConfig);
-    }
 }