+++ app/src/main/java/ly/count/android/demo/ActivityExampleOthers.java
@@ -141,22 +141,4 @@ public class ActivityExampleOthers extends AppCompatActivity {
     public void onClickEndSession(View v) {
         Countly.sharedInstance().sessions().endSession();
     }
-
-    @Override
-    public void onStart() {
-        super.onStart();
-        Countly.sharedInstance().onStart(this);
-    }
-
-    @Override
-    public void onStop() {
-        Countly.sharedInstance().onStop();
-        super.onStop();
-    }
-
-    @Override
-    public void onConfigurationChanged(Configuration newConfig) {
-        super.onConfigurationChanged(newConfig);
-        Countly.sharedInstance().onConfigurationChanged(newConfig);
-    }
 }