+++ app/src/main/java/ly/count/android/demo/MainActivity.java
@@ -141,22 +141,4 @@ public class MainActivity extends AppCompatActivity {
     public void onClickButtonRatings(View v) {
         startActivity(new Intent(this, ActivityExampleFeedback.class));
     }
-
-    @Override
-    public void onStart() {
-        super.onStart();
-        Countly.sharedInstance().onStart(this);
-    }
-
-    @Override
-    public void onStop() {
-        Countly.sharedInstance().onStop();
-        super.onStop();
-    }
-
-    @Override
-    public void onConfigurationChanged(Configuration newConfig) {
-        super.onConfigurationChanged(newConfig);
-        Countly.sharedInstance().onConfigurationChanged(newConfig);
-    }
 }